Create SoundArduino.ino
This is the final code for the Sound Arduino. This Arduino is tasked with communicating with the speaker module using the TMRpcm library (allows for asynchronous playback) and additionally receives its commands from the Rate Arduino. This Arduino also makes use of the SD card library and can play its audio from an attached SD card reader.